move connection creation into connection interface
