+++ src/model/networking/server/server-connection.cc
@@ -24,11 +24,6 @@ using namespace model_message_functionality;
 using namespace model_networking_utility;
 using json = nlohmann::json;
 
-// get sockaddr, IPv4 or IPv6:
-void *ServerConnection::GetInAddr(struct sockaddr *sa) {
-  return &(((struct sockaddr_in6 *)sa)->sin6_addr);
-}
-
 void ServerConnection::SetState(SocketHandler *next_handler) {
   delete socket_handler;
   socket_handler = next_handler;
@@ -47,54 +42,6 @@ ServerConnection::ServerConnection(const std::string &ip_address,
   this->sockfd = -1;
   this->stream_out_factory = std::make_shared<NetworkStreamOutFactory>();
   this->stream_in_factory = std::make_shared<NetworkStreamInFactory>();
-}
-
-int ServerConnection::CreateConnection() {
-  struct addrinfo hints, *servinfo, *p;
-  int rv;
-  char s[INET6_ADDRSTRLEN];
-
-  memset(&hints, 0, sizeof hints);
-  hints.ai_family = AF_INET6;  // add a 6 to make it work with ivp 6 - remove 6 for ivp 4
-  hints.ai_socktype = SOCK_STREAM;
-  hints.ai_protocol = IPPROTO_TCP;
-
-  if ((rv = getaddrinfo(ip_address.c_str(), port.c_str(), &hints, &servinfo)) !=
-      0) {
-    fprintf(stderr, "getaddrinfo: %s\n", gai_strerror(rv));
-    return -1;
-  }
-
-  // loop through all the results and connect to the first we can
-  for (p = servinfo; p != NULL; p = p->ai_next) {
-    if ((sockfd = socket(p->ai_family, p->ai_socktype, p->ai_protocol)) == -1) {
-      perror("client: socket");
-      continue;
-    }
-
-    if (connect(sockfd, p->ai_addr, p->ai_addrlen) == -1) {
-      close(sockfd);
-      perror("client: connect");
-      continue;
-    }
-
-    break;
-  }
-
-  if (p == NULL) {
-    fprintf(stderr, "client: failed to connect\n");
-    return -1;
-  }
-
-  inet_ntop(p->ai_family, GetInAddr((struct sockaddr *)p->ai_addr), s,
-            sizeof s);
-  printf("client: connecting to %s\n", s);
-
-  freeaddrinfo(servinfo);  // all done with this structure
-
-  SetState(new InsecureSocketHandler(sockfd));
-
-  return sockfd;
 }
 
 int ServerConnection::SendPublicKey() {
